+++ mex/common/primary_source/extract.py
@@ -1,7 +1,7 @@
 import json
 from functools import lru_cache
-from pathlib import Path
 
+from mex.common.assets.registry import get_assets_connector
 from mex.common.logging import logger
 from mex.common.primary_source.models import SeedPrimarySource
 from mex.common.settings import BaseSettings
@@ -18,7 +18,7 @@ def extract_seed_primary_sources() -> list[SeedPrimarySource]:
         List of seed primary sources
     """
     settings = BaseSettings.get()
-    with Path(settings.primary_sources_path).open() as fh:
-        raw_units = json.load(fh)
+    connector = get_assets_connector()
+    raw_units = json.loads(connector.read(settings.primary_sources_path))
     logger.info("extracted %s seed primary sources", len(raw_units))
     return [SeedPrimarySource.model_validate(raw) for raw in raw_units]